array: Introduce inner array representation parameter
This is almost an entire rewrite of the existing array data structure.
This way, we are able to use a single type for arrays owning their data
(A = Vec<T>), array views (A = &[T]) and mutable array views (A = &mut [T]).

The operations we implement on arrays, e.g. Add, Sub or Mul benefit from
this additional parameter since the caller can just call view() on the
right hand side before invoking the operation. This way, the array is not
consumed.

In many cases, this facilitates zero-copy array operations where one had to
clone lots of data before, all while keeping a single, unified API.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>