stat: Add standardize() functions
By default, standardize2() normalizes each row, calculating
the mean and standard deviation for that particular row.

When working with ML datasets, we often have input data which is
organized in a way such that each row is an observation and the columns
are the observed features. In that case, we usually want to standardize
each column instead. This can be achieved by using axis = 1.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>